baya bi change

+++ GTBack/DTO/Randevu/Request/CreateCompanyDTO.cs
@@ -4,7 +4,7 @@ public class CreateCompanyDTO
 {
     public String Name { get; set; }
     public String Address { get; set; }
-    public int CompanyTypeId { get; set; }
+    public long CompanyTypeId { get; set; }
     public String Mail { get; set; }
     public String Phone { get; set; }
     public float Latitude { get; set; }
