+++ GTBack/DTO/Randevu/Request/EventAddRequestDTO.cs
@@ -5,9 +5,9 @@ public class EventAddRequestDTO
     public DateTime StartDateTime { get; set; }
     public DateTime EndDateTime { get; set; }
     public String Description { get; set; }
-    public int CurrencyId { get; set; }
+    public long CurrencyId { get; set; }
     public int StatusId { get; set; }
-    public int AdminUserId { get; set; }
-    public int ClientUserId { get; set; }
-    public int EventTypeId { get; set; }
+    public long AdminUserId { get; set; }
+    public long ClientUserId { get; set; }
+    public long EventTypeId { get; set; }
 }
