+++ GTBack/DTO/Randevu/Request/UserRegisterDTO.cs
@@ -3,7 +3,7 @@ namespace GTBack.Core.DTO;
 public class UserRegisterDTO:BaseRegisterDTO
 {
 
-    public int UserTypeId { get; set; }
-    public int CompanyId { get; set; }
+    public long UserTypeId { get; set; }
+    public long CompanyId { get; set; }
     public string Password { get; set; }
 }
