+++ GTBack/DTO/Randevu/Response/EventByEventId.cs
@@ -5,11 +5,11 @@ public class EventByEventId
     public DateTime StartDateTime { get; set; }
     public DateTime EndDateTime { get; set; }
     public String Description { get; set; }
-    public int AdminUserId { get; set; }
+    public long AdminUserId { get; set; }
     public UserDTO Client { get; set; }
-    public int ClientUserId { get; set; }
+    public long ClientUserId { get; set; }
     public EventTypeDTO EventTypeDto { get; set; }
     public int StatusId { get; set; }
     public int Price { get; set; }
-    public int Id { get; set; }
+    public long Id { get; set; }
 }
