+++ GTBack/DTO/Randevu/Response/EventTypeDTO.cs
@@ -2,7 +2,7 @@ namespace GTBack.Core.DTO;
 
 public class EventTypeDTO
 {
-    public int EventTypeId { get; set; }
+    public long EventTypeId { get; set; }
     public string EventName { get; set; }
     public int Duration{ get; set; }
     public string Description{ get; set; }
