+++ GTBack/Entities/Randevu/EventTypeCompanyRelation.cs
@@ -1,10 +1,12 @@
+using GTBack.Core.Entities.SharedEntities;
+
 namespace GTBack.Core.Entities;
 
 public class EventTypeCompanyRelation:BaseEntity
 {
  
-    public int EventTypeId { get; set; }
-    public int CompanyId { get; set; }
+    public long EventTypeId { get; set; }
+    public long CompanyId { get; set; }
     public virtual EventType EventType{ get; set; }
     public virtual Company Company { get; set; }
