+++ GTBack/Entities/Randevu/FAQ.cs
@@ -1,6 +1,7 @@
 
 
 using System.ComponentModel.DataAnnotations.Schema;
+using GTBack.Core.Entities.SharedEntities;
 
 namespace GTBack.Core.Entities;
 
@@ -9,10 +10,10 @@ public class FAQ:BaseEntity
     public String Question { get; set; }
     public String Answer { get; set; }
     public int Like { get; set; }
-    public int CompanyId { get; set; }
+    public long CompanyId { get; set; }
     
-    public int SenderUserId { get; set; }
-    public int? AnsweredUserId { get; set; }
+    public long SenderUserId { get; set; }
+    public long? AnsweredUserId { get; set; }
     public  User SenderUser { get; set; }
     public  User? AnsweredUser { get; set; }
     public  Company Company { get; set; }
