+++ GTBack/Entities/Randevu/User.cs
@@ -1,3 +1,5 @@
+using GTBack.Core.Entities.SharedEntities;
+
 namespace GTBack.Core.Entities;
 
 public class User  : BaseEntity
@@ -9,9 +11,9 @@ public class User  : BaseEntity
     public String Phone { get; set; }
     public String Address { get; set; }
     public string PasswordHash { get; set; }
-    public int UserTypeId { get; set; }
+    public long UserTypeId { get; set; }
     
-    public int? CompanyId { get; set; }
+    public long? CompanyId { get; set; }
     
     public  Company Company { get; set; }
