+++ GTBack/Entities/Restourant/Order.cs
@@ -4,8 +4,8 @@ public class Order:BaseEntity
 {
     public String Name { get; set; }
     public String OrderNote { get; set; }
-    public int? ExtraMenuItemId { get; set; }
-    public int AdditionId { get; set; }
+    public long? ExtraMenuItemId { get; set; }
+    public long AdditionId { get; set; }
     public int OrderStatus { get; set; }
     public int OrderStartDate { get; set; }
     public int OrderDeliveredDate { get; set; }
