+++ GTBack/Entities/Restourant/Payment.cs
@@ -2,7 +2,7 @@ namespace GTBack.Core.Entities.Restourant;
 
 public class Payment:BaseEntity
 {
-    public  int AdditionId { get; set; }
+    public  long AdditionId { get; set; }
     public  int AmountPaid { get; set; }
     public  int PaymentMethod { get; set; }
     public  DateTime PaymentDate { get; set; }
